refactor(settings): centralize Claude auth process helpers
